Move stat view implementations into a submodule
Also move serialize_distributed_ddls into commands, seems like an oversight from
last year (by me).
