+++ src/backend/distributed/test/citus_stat_tenants.c
@@ -15,7 +15,7 @@
 
 #include "sys/time.h"
 
-#include "distributed/utils/citus_stat_tenants.h"
+#include "distributed/stats/stat_tenants.h"
 
 PG_FUNCTION_INFO_V1(sleep_until_next_period);
